[CV-16] Simplify logger
