+++ common/src/main/java/net/crystalixs/core/common/config/ConfigChangeLogger.java
@@ -2,8 +2,14 @@ package net.crystalixs.core.common.config;
 
 import java.util.logging.Logger;
 
-public interface ConfigChangeLogger {
+public final class ConfigChangeLogger {
 
-    void log(Logger logger, String fileName, ConfigChangeSet changeSet);
+    public void log(Logger logger, String fileName, ConfigChangeSet changeSet) {
+        if (!changeSet.hasChanges()) return;
 
+        logger.info("Synchronized config " + fileName + ":");
+        for (String entry : changeSet.getEntries()) {
+            logger.info("  " + entry);
+        }
+    }
 }
