--- common/src/main/java/net/crystalixs/core/common/config/DefaultConfigChangeLogger.java
@@ -1,16 +1,0 @@
-package net.crystalixs.core.common.config;
-
-import java.util.logging.Logger;
-
-public class DefaultConfigChangeLogger implements ConfigChangeLogger {
-
-    @Override
-    public void log(Logger logger, String fileName, ConfigChangeSet changeSet) {
-        if (!changeSet.hasChanges()) return;
-
-        logger.info("Synchronized config " + fileName + ":");
-        for (String entry : changeSet.getEntries()) {
-            logger.info("  " + entry);
-        }
-    }
-}
